requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
streamlit
langchain
langchain-groq
requests
python-dotenv
